구조 변경 중간 커밋

+++ src/main/java/com/website/trip/biz/dao/BaseDao.java
@@ -12,6 +12,6 @@ public interface BaseDao<T extends BaseDto> {
 
     T selectOne(T parameter);
     List<T> selectList(T module);
-    int selectListCount(T module);
+    long selectListCount(T module);
 
 }
